% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADICohortBuild.R
\name{ADICohortBuild}
\alias{ADICohortBuild}
\title{ADICohortBuild}
\usage{
ADICohortBuild(
  connection,
  cohortDatabaseSchema,
  cdmDatabaseSchema,
  cohortTable,
  outputFolder,
  minCellCount = 10
)
}
\arguments{
\item{connection}{An object of type \code{connection} that has an established database connection.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{minCellCount}{The minimum number of subjects contributing to a count before it can be included
                            in packaged results.

table examples

1. cohort_table
-----------------------------------------------------------------------
cohort_definition_id | subject_id | cohort_start_date | cohort_end_date
1775946              | 1000001    | 2000-04-21        | 2000-08-22
-----------------------------------------------------------------------

2. location (*latitude and longitude will be available in OMOP CDM 5.4)
--------------------------------------------------------------------------------------------------------------------------------------
location_id | address_1 | address_2 | city | state | zip | county | location_source_value | latitude | longitude
1       | 36 BAY STREET | 1A        | BRONX| NY    | 10468|       |                       | 40.815197| -72.509873
--------------------------------------------------------------------------------------------------------------------------------------

3. address_geocoded (*this is a subset of the geocoded table which should be created using local geocoding services, e.g. ArcGIS, OHDSI GIS)
--------------------------------------------------------------------------------------------------------------------------------------
location_id | FIPS_BG_ID   | STATE_NAME | COUNTY_NAME
1           | 360595183004 | NY         | BRONX
--------------------------------------------------------------------------------------------------------------------------------------}
}
\description{
ADICohortBuild
}
\details{
Run the patient count and median ADI for each County
}
