% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateAndLoadFileToTable.R
\name{createAndLoadFileToTable}
\alias{createAndLoadFileToTable}
\title{CreateAndLoadFileToTable
Execute the package to load ADI into the cohort database and link data with your cohort result}
\usage{
createAndLoadFileToTable(
  pathToCsv,
  sep = ",",
  connection,
  cohortDatabaseSchema,
  createTableFile,
  tableName,
  targetDialect = "sql server",
  oracleTempSchema,
  package
)
}
\arguments{
\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that hold the cohorts created from previous studies.
which will be linked with ADI data.}
}
\description{
CreateAndLoadFileToTable
Execute the package to load ADI into the cohort database and link data with your cohort result
}
