% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Main}
\usage{
execute(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  oracleTempSchema,
  outputFolder,
  loadADI = TRUE,
  geocodedLocation = TRUE,
  geocodedLocationTable,
  createViz = TRUE,
  minCellCount = 10
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that hold the cohorts created from previous studies.
which will be linked with ADI data.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{loadADI}{Does ADI data need to be loaded from the package into database?}

\item{geocodedLocation}{Is your location data geocoded with Census Block Group FIPS available to run ADI?}

\item{geocodedLocationTable}{The name of the table that hold the geocoded location and Census Block Group FIPS.}

\item{createViz}{Generate map visualizations through package?}

\item{minCellCount}{The minimum number of subjects contributing to a count before it can be shown
on the maps.}
}
\description{
Main
}
\details{
This function executes the ADIOMOPLinkage Study. Execute the package to load ADI into the cohort database and link data with your cohort result
}
